pyproject.toml
[project]
name = "frontier"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk",
    "beautifulsoup4",
    "fastapi",
    "joblib>=1.5.1",
    "langchain",
    "langchain-chroma",
    "langchain-community",
    "langchain-docling",
    "langchain-huggingface>=0.2.0",
    "langchain-mcp-adapters",
    "langchain-openai",
    "langgraph",
    "langmem>=0.0.27",
    "matplotlib>=3.10.3",
    "mcp-server-sqlite",
    "mcp[cli]",
    "mdit-py-plugins",
    "nb-cli",
    "nonebot-adapter-onebot",
    "nonebot-adapter-qq",
    "nonebot2",
    "openai",
    "pandas-stubs>=2.2.3.250527",
    "pillow",
    "playwright",
    "pypinyin",
    "ruff>=0.11.12",
    "scikit-learn>=1.6.1",
    "sentence-transformers>=4.1.0",
    "tavily-python",
    "websockets",
]
[tool.nonebot]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" },
    { name = "QQ", module_name = "nonebot.adapters.qq" }
]
plugins = []
plugin_dirs = ["plugins"]
builtin_plugins = []

[tool.uv.workspace]
members = [
    "plugins/multi_agent_a2a",
]
[tool.black]
line-length = 119
target-version = ['py312']

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]
extend-include = ["*.ipynb"]

# Same as Black.
line-length = 119
indent-width = 4

# Assume Python 3.8
target-version = "py312"

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
# McCabe complexity (`C901`) by default.
select = ["E4", "E7", "E9", "F"]
ignore = []

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

# Enable auto-formatting of code examples in docstrings. Markdown,
# reStructuredText code/literal blocks and doctests are all supported.
#
# This is currently disabled by default, but it is planned for this
# to be opt-out in the future.
docstring-code-format = true

# Set the line length limit used when formatting code snippets in
# docstrings.
#
# This only has an effect when the `docstring-code-format` setting is
# enabled.
docstring-code-line-length = "dynamic"
